fix u64/i64 typedefs (long -> long long)

+++ RGFW/RGFW.h
@@ -165,8 +165,8 @@ extern "C" {
 		typedef signed short 	i16;
 		typedef unsigned int 	u32;
 		typedef signed int		i32;
-		typedef unsigned long	u64;
-		typedef signed long		i64;
+		typedef unsigned long long	u64;
+		typedef signed long long		i64;
 	#else
 		#include <stdint.h>
